+++ dionysos-files-mode.el
@@ -27,6 +27,7 @@
 
 (require 'dionysos-backend)
 (require 'dionysos-mode)
+(require 'dionysos-volume)
 
 
 ;; Management
@@ -81,8 +82,8 @@
     (define-key map (kbd "q") 'dionysos-files-mode-quit)
     (define-key map (kbd "p") 'dionysos-files-mode-previous)
     (define-key map (kbd "n") 'dionysos-files-mode-next)
-    (define-key map (kbd "+") 'dionysos--volume-raise)
-    (define-key map (kbd "-") 'dionysos--volume-decrease)
+    (define-key map (kbd "+") 'dionysos-volume-raise)
+    (define-key map (kbd "-") 'dionysos-volume-decrease)
     (define-key map (kbd "RET") 'dionysos-files-mode-start)
     (define-key map (kbd "SPC") 'dionysos-files-mode-stop)
     map)