+++ ChangeLog.md
@@ -1,10 +1,14 @@
 # dionysos ChangeLog
 
-# Version 0.2.0 (IN PROGRESS)
+# Version 0.3.0 (18/03/2015)
 
-- Manage volume
+- Add MPD mode to display current playlist
+- Add MPD backend
+
+# Version 0.2.0 (03/02/2015)
+
+- [GH-6][]: Manage volume
 - [GH-2][]: MPlayer backend
-
 
 # Version 0.1.0 (15/01/2015)
 
@@ -20,3 +24,4 @@
 
 [GH-1]: https://github.com/nlamirault/dionysos/issues/1
 [GH-2]: https://github.com/nlamirault/dionysos/issues/2
+[GH-6]: https://github.com/nlamirault/dionysos/issues/6